Move orderedPairs to util
Might come in handy elsewhere one day...
